libsnk: add snk configuration with joint buffer wrap parameter
The configuration controls game's behaviour. The wrap_joints
parameter mitigates joint buffer overflow by overwriting the
oldest entries.

That makes so that game does not stop on that condition, but the
length of a snake (as well as the player's score) reduces by
the size of the last joint that was overwritten.